refactor(xrpl): memo tx field type
